fix error on usr_trb1
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@2416 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: 3561df2d54f13a44b9c011b7db1c68170386e436
Former-commit-id: 776ee308924b21c0e340f0d29de5c2e655a637fb

+++ src/drivers/usr_trb1/driver.cpp
@@ -3335,7 +3335,7 @@ bool Driver::isStuck()
          (car->_trkPos.toRight < 1.0 && car->_trkPos.toMiddle < rldata->offset - 2 && angle < -0.7)))
     {
       // yes, we're stuck
-      stuck_timer = (float)last_stuck_time = (float)simtime;
+      stuck_timer = last_stuck_time = (float)simtime;
       setMode( mode_correcting );
 
       stuck = STUCK_REVERSE;
